+++ src/index.ts
@@ -2,6 +2,6 @@ export * from "./adapter/index.js";
 export * from "./http/index.js";
 export * from "./middleware/index.js";
 export * from "./authentication/index.js";
-export * from "./task/index.js";
+export * from "./tasks/index.js";
 export * from "./utils/Constants.js";
 export * from "./utils/Version.js";
