+++ src/tasks/UploadResult.ts
