+++ src/tasks/UploadSliceRequestBuilder.ts
